user cannot delete others msgs + after send keyboard hide + saving user with token

+++ lib/controller/cloud_controller.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:chat_app/controller/auth_controller.dart';
 import 'package:chat_app/model/chat_model.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 
 late FirebaseFirestore cloud;
@@ -12,9 +13,10 @@ class Cloud extends AuthController {
     cloud = FirebaseFirestore.instance;
   }
 
-  storeToken(String token) async {
+  storeToken(String token, String name) async {
     String uid = getUserId;
     await cloud.collection('users').doc(uid).set({
+      'name': name,
       'token': FieldValue.arrayUnion([token])
     });
   }
@@ -23,13 +25,20 @@ class Cloud extends AuthController {
     await cloud.collection('chats').add(chat.toMap());
   }
 
-  deleteChat(Timestamp time) async {
-    var get =
-        await cloud.collection('chats').where('time', isEqualTo: time).get();
-    await get.docs.first.reference.delete();
+  Future<dynamic> deleteChat(Timestamp time, Chat chat, User? user) async {
+    if (chat.uid == user!.uid) {
+      var get =
+          await cloud.collection('chats').where('time', isEqualTo: time).get();
+      return await get.docs.first.reference.delete();
+    } else {
+      return 'null';
+    }
   }
 
   Stream<QuerySnapshot<Map<String, dynamic>>> loadChat() {
-    return cloud.collection('chats').snapshots();
+    return cloud
+        .collection('chats')
+        .orderBy('time', descending: true)
+        .snapshots();
   }
 }
